Changed name to ...DoubleFilterTests since it was testing just methods from that class
